Separate Swift tests into trait
